requirements.txt
langchain==0.2.2
langchain-community==0.2.3
jq==1.7.0
python-dotenv==1.0.1
langchain-openai==0.1.8
faiss-cpu==1.8.0
flask==3.0.3
uuid==1.30
gunicorn==22.0.0